+++ server/templates/@client=sapper/src/routes/blog/index.svelte
@@ -1,4 +1,6 @@
 <script context="module" lang="ts">
+	import { pagesPath } from "$path"
+
 	export function preload() {
 		return this.fetch(`blog.json`).then((r: { json: () => any; }) => r.json()).then((posts: { slug: string; title: string, html: any }[]) => {
 			return { posts };
@@ -29,6 +31,6 @@
 				tell Sapper to load the data for the page as soon as
 				the user hovers over the link or taps it, instead of
 				waiting for the 'click' event -->
-		<li><a rel="prefetch" href="blog/{post.slug}">{post.title}</a></li>
+		<li><a rel="prefetch" href={pagesPath.blog._slug(post.slug).$url()}>{post.title}</a></li>
 	{/each}
 </ul>