+++ server/templates/server/api/article/index.ts
@@ -1,0 +1,8 @@
+import type { ArticleInfo } from '$/service/article'
+
+export type Methods = {
+  get: {
+    query: { search?: string }
+    resBody: ArticleInfo[]
+  }
+}